# OpenApiSpec.yml
openapi: "3.0.0"
info:
  title: "express-api"
  version: "1.2.4"
servers:
  - url: "https://localhost/api/v1"
    description: "Local development server"
  - url: "http://localhost:8888/api/v1"
    description: "Unsecured Local development server"
  - url: "https://api.example.com/api/v1"
    description: "Production server"
paths:
...
  /groups/{groupID}:
    get:
      description: "Get a group by ID"
      tags:
        - "Groups"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "groupID"
          required: true
          schema:
            type: "string"
          description: "ID of the group to retrieve"
        - in: "query"
          name: "with"
          schema:
            type: "string"
            enum:
              - "users"
          description: "Include related data (users) in the response"
      responses:
        "200":
          description: "Returns the group information"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "string"
                  name:
                    type: "string"
                  users:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/User"
                    description: "Only included when with=users query parameter is provided"
    post:
      description: "Update a group's information"
      tags:
        - "Groups"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "groupID"
          required: true
          schema:
            type: "string"
          description: "ID of the group to update"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  description: "New name for the group"
      responses:
        "200":
          description: "Returns the updated group"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupID}/users/invite:
    post:
      description: "Invite a user to a group"
      tags:
        - "Groups"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "groupID"
          required: true
          schema:
            type: "string"
          description: "ID of the group to invite the user to"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                email:
                  type: "string"
                  format: "email"
                  description: "Email of the user to invite"
                role:
                  type: "string"
                  enum:
                    - "User"
                    - "Admin"
                  default: "User"
                  description: "Role to assign to the user in the group"
      responses:
        "200":
          description: "User successfully invited to the group"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  groupID:
                    type: "string"
                  userID:
                    type: "string"
                  role:
                    type: "string"
                    enum:
                      - "User"
                      - "Admin"
  /groups/{groupID}/users/{userID}/resend-invitation-email:
    post:
      tags:
        - "Groups"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "groupID"
          required: true
          schema:
            type: "string"
          description: "ID of the group"
        - in: "path"
          name: "userID"
          required: true
          schema:
            type: "string"
          description: "ID of the user to resend invitation to"
      responses:
        "200":
          description: "Invitation email resent successfully"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  email:
                    type: "string"
                    description: "Email address of the user"
  /groups/{groupID}/users/{userID}/set-role:
    post:
      description: "Set a user's role within a group"
      tags:
        - "Groups"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "groupID"
          required: true
          schema:
            type: "string"
          description: "ID of the group"
        - in: "path"
          name: "userID"
          required: true
          schema:
            type: "string"
          description: "ID of the user whose role will be changed"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                role:
                  type: "string"
                  enum:
                    - "User"
                    - "Admin"
                  default: "User"
                  description: "The role to assign to the user"
      responses:
        "200":
          description: "Role updated successfully"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  userID:
                    type: "string"
                  groupID:
                    type: "string"
                  role:
                    type: "string"
                    enum:
                      - "User"
                      - "Admin"
  /groups/{groupID}/users/{userID}:
    delete:
      description: "Remove a user from a group"
      tags:
        - "Groups"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "groupID"
          required: true
          schema:
            type: "string"
          description: "ID of the group"
        - in: "path"
          name: "userID"
          required: true
          schema:
            type: "string"
          description: "ID of the user to remove from the group"
      responses:
        "200":
          description: "User successfully removed from the group"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  groupID:
                    type: "string"
                  userID:
                    type: "string"
components:
  schemas:
...
    Group:
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        ownerID:
          type: "string"
          format: "uuid"
        deletedAt:
          type: "string"
          format: "date-time"
          nullable: true
    User:
      properties:
        id:
          type: "string"
          format: "uuid"
        email:
          type: "string"
          format: "email"
        firstName:
          type: "string"
        lastName:
          type: "string"
        bio:
          type: "string"
        emailVerified:
          type: "boolean"
        lastLoginAt:
          type: "string"
          format: "date-time"
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
...