# src/routes/Groups.ts
import { body, validationResult, matchedData, param } from 'express-validator';
import { generateEmail } from './../providers/Helpers';
import { GroupUser } from './../models/GroupUser';
import passport from './../providers/Passport';
import { Group } from './../models/Group';
import middleware from './middleware';
import User from './../models/User';
import bcrypt from 'bcryptjs';
import express from 'express';
import crypto from 'crypto';

export const app = express.Router();

/**
 * @swagger
 * tags:
 *   - name: Groups
 *     description: Groups
 */

/**
 * @openapi
 * /groups/{groupID}:
 *   get:
 *     description: Get a group by ID
 *     tags: [Groups]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: groupID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the group to retrieve
 *       - in: query
 *         name: with
 *         schema:
 *           type: string
 *           enum: [users]
 *         description: Include related data (users) in the response
 *     responses:
 *       200:
 *         description: Returns the group information
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 id:
 *                   type: string
 *                 name:
 *                   type: string
 *                 users:
 *                   type: array
 *                   items:
 *                     $ref: '#/components/schemas/User'
 *                   description: Only included when with=users query parameter is provided
 */
app.get('/groups/:groupID', [
    passport.authenticate('jwt', { session: false }),
    param('groupID').exists().isUUID(),
    middleware.isInGroup,
], async (req: express.Request, res: express.Response, next: express.NextFunction) => {
    try {
        const { groupID } = matchedData(req);
        return res.json(await Group.findByPk(groupID, {
            include: (req.query.with === 'users') ? [User] : [],
            rejectOnEmpty: true,
        }));
    } catch (error) {
        return next(error);
    }
});

/**
 * @openapi
 * /groups/{groupID}:
 *   post:
 *     description: Update a group's information
 *     tags: [Groups]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: groupID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the group to update
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               name:
 *                 type: string
 *                 description: New name for the group
 *     responses:
 *       200:
 *         description: Returns the updated group
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Group'
 */
app.post('/groups/:groupID', [
    passport.authenticate('jwt', { session: false }),
    param('groupID').isUUID(),
    body('name').exists().isString(),
    middleware.hasRole('Admin'),
], async (req: express.Request, res: express.Response, next: express.NextFunction) => {
    try {
        const errors = validationResult(req);
        if (!errors.isEmpty()) return res.status(422).json({ errors: errors.mapped() });
        const { groupID, name } = matchedData(req);

        const group = await Group.findByPk(groupID, { rejectOnEmpty: true });
        await group.update({ name });
        return res.json(group);
    } catch (error) {
        return next(error);
    }
});

/**
 * @openapi
 * /groups/{groupID}/users/invite:
 *   post:
 *     description: Invite a user to a group
 *     tags: [Groups]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: groupID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the group to invite the user to
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               email:
 *                 type: string
 *                 format: email
 *                 description: Email of the user to invite
 *               role:
 *                 type: string
 *                 enum: [User, Admin]
 *                 default: User
 *                 description: Role to assign to the user in the group
 *     responses:
 *       200:
 *         description: User successfully invited to the group
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 groupID:
 *                   type: string
 *                 userID:
 *                   type: string
 *                 role:
 *                   type: string
 *                   enum: [User, Admin]
 */
app.post('/groups/:groupID/users/invite', [
    passport.authenticate('jwt', { session: false }),
    param('groupID').isUUID(),
    body('email').isEmail().toLowerCase(),
    body('role').default('User').isIn(['User', 'Admin']),
    middleware.hasRole('Admin'),
], async (req: express.Request, res: express.Response, next: express.NextFunction) => {
    try {
        const errors = validationResult(req);
        if (!errors.isEmpty()) return res.status(422).json({ errors: errors.mapped() });
        const { email, role, groupID } = matchedData(req);

        const group = await Group.findByPk(groupID, { rejectOnEmpty: true });

        let user = await User.findOne({
            where: { email },
        });

        if (user) {
            if (user.id === req.user.id) return res.status(401).json({
                msg: 'You cannot invite yourself to a group',
                code: 98644,
            });

            // Check if relationship already exists
            const relationship = await GroupUser.findOne({
                where: {
                    groupID,
                    userID: user.id,
                },
            });

            if (relationship) {
                return res.status(401).json({
                    msg: 'User is already a member of this group',
                    code: 98645,
                });
            } else {
                await GroupUser.create({
                    groupID,
                    userID: user.id,
                    role,
                });
            }
        } else {
            user = await User.create({
                email,
                password: bcrypt.hashSync(crypto.randomBytes(20).toString('hex'), bcrypt.genSaltSync(10)), // AB: Random password, will be updated by user
                firstName: '',
                emailVerificationKey: String(Math.floor(Math.random() * (999999 - 111111 + 1)) + 111111),
                tos: null,
            });

            const relationship = await GroupUser.create({
                groupID,
                userID: user.id,
                role,
                inviteKey: crypto.randomBytes(10).toString('hex'),
            });

            //////////////////////////////////////////
            // EMAIL THIS TO THE USER
            const link = `${process.env.FRONTEND_URL}/invite/${relationship.inviteKey}`;
            if (typeof global.it !== 'function') console.log(`\n\nINVITE LINK: ${link}\n\n`);
            // const html = generateEmail('Default', { 
            //     title: `You've been invited to join ${group.name}`,
            //     body: `You have been invited to join ${group.name} by ${req.user.firstName} ${req.user.lastName}. Click the button below to accept the invitation.`,
            //     link,
            //     linkText: 'Accept Invitation'
            // });
            //////////////////////////////////////////
        }

        return res.json({
            groupID,
            userID: user.id,
            role,
        });
    } catch (error) {
        return next(error);
    }
});

/**
 * @swagger
 * /groups/{groupID}/users/{userID}/resend-invitation-email:
 *   post:
 *     tags:
 *       - Groups
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: groupID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the group
 *       - in: path
 *         name: userID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the user to resend invitation to
 *     responses:
 *       200:
 *         description: Invitation email resent successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 email:
 *                   type: string
 *                   description: Email address of the user
 */
app.post('/groups/:groupID/users/:userID/resend-invitation-email', [
    passport.authenticate('jwt', { session: false }),
    param('groupID').isUUID(),
    param('userID').isUUID(),
    middleware.hasRole('Admin'),
], async (req: express.Request, res: express.Response, next: express.NextFunction) => {
    try {
        const errors = validationResult(req);
        if (!errors.isEmpty()) return res.status(422).json({ errors: errors.mapped() });
        const { userID, groupID } = matchedData(req);

        const relationship = await GroupUser.unscoped().findOne({
            where: { userID, groupID },
        });

        if (!relationship) return res.status(400).json({ msg: 'Invite not found' });

        const user = await User.unscoped().findByPk(relationship.userID, { rejectOnEmpty: true });
        const group = await Group.findByPk(relationship.groupID, { rejectOnEmpty: true });

        if (!relationship.inviteKey) return res.status(400).json({
            msg: 'This user has already accepted their invitation',
            code: 40422,
        });

        await relationship.update({
            inviteKey: crypto.randomBytes(10).toString('hex'),
        });

        //////////////////////////////////////////
        // EMAIL THIS TO THE USER
        const link = `${process.env.FRONTEND_URL}/invite/${relationship.inviteKey}`;
        if (typeof global.it !== 'function') console.log(`\n\nINVITE LINK: ${link}\n\n`);
        // const html = generateEmail('Invite', { link, groupName: group.name });
        //////////////////////////////////////////

        return res.json({ email: user.email });
    } catch (error) {
        return next(error);
    }
});

/**
 * @openapi
 * /groups/{groupID}/users/{userID}/set-role:
 *   post:
 *     description: Set a user's role within a group
 *     tags: [Groups]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: groupID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the group
 *       - in: path
 *         name: userID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the user whose role will be changed
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               role:
 *                 type: string
 *                 enum: [User, Admin]
 *                 default: User
 *                 description: The role to assign to the user
 *     responses:
 *       200:
 *         description: Role updated successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 userID:
 *                   type: string
 *                 groupID:
 *                   type: string
 *                 role:
 *                   type: string
 *                   enum: [User, Admin]
 */
app.post('/groups/:groupID/users/:userID/set-role', [
    passport.authenticate('jwt', { session: false }),
    param('groupID').isUUID(),
    param('userID').isUUID(),
    body('role').default('User').isIn(['User', 'Admin']),
    middleware.hasRole('Admin'),
], async (req: express.Request, res: express.Response, next: express.NextFunction) => {
    try {
        const errors = validationResult(req);
        if (!errors.isEmpty()) return res.status(422).json({ errors: errors.mapped() });
        const { role, groupID, userID } = matchedData(req);

        // Check if user exists
        const user = await User.findByPk(userID, { rejectOnEmpty: true });

        const group = await Group.findByPk(groupID, { rejectOnEmpty: true });

        if (group.ownerID === userID) return res.status(400).json({
            msg: 'You cannot remove the owner of the group',
            code: 20330,
        });

        const groupUsers = await GroupUser.findOne({
            where: {
                groupID, userID,
            },
        });

        if (!groupUsers) return res.status(400).json({
            msg: `User ${userID} does not exist in group ${groupID}`,
            code: 10424,
        });

        await groupUsers.update({ role });

        return res.json({ userID, groupID, role });
    } catch (error) {
        return next(error);
    }
});

/**
 * @openapi
 * /groups/{groupID}/users/{userID}:
 *   delete:
 *     description: Remove a user from a group
 *     tags: [Groups]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: groupID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the group
 *       - in: path
 *         name: userID
 *         required: true
 *         schema:
 *           type: string
 *         description: ID of the user to remove from the group
 *     responses:
 *       200:
 *         description: User successfully removed from the group
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 groupID:
 *                   type: string
 *                 userID:
 *                   type: string
 */
app.delete('/groups/:groupID/users/:userID', [
    passport.authenticate('jwt', { session: false }),

    param('groupID').isUUID(),
    param('userID').isUUID(),

    middleware.isNotSelf,
    middleware.hasRole('Admin'),
], async (req: express.Request, res: express.Response, next: express.NextFunction) => {
    try {
        const errors = validationResult(req);
        if (!errors.isEmpty()) return res.status(422).json({ errors: errors.mapped() });
        const { groupID, userID } = matchedData(req);

        const group = await Group.findByPk(groupID, { rejectOnEmpty: true });

        if (group.ownerID === userID) return res.status(400).json({
            msg: 'You cannot remove the owner of the group',
            code: 63930,
        });

        await GroupUser.destroy({
            where: {
                groupID, userID,
            },
        });

        return res.json({ userID, groupID });
    } catch (error) {
        return next(error);
    }
});